pyproject.toml
[tool.poetry]
name = "name-tags-generator"
version = "0.1.0"
description = ""
authors = ["amit rahav <amit.r.89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.1.0"
click = "^8.1.2"
pandas = "^1.4.2"
openpyxl = "^3.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
